DT504: Add `updateFont` JS to handle replacing the IDML font
In this commit, all projects are updated to use "Adobe Arabic".

The `updateFont` JS is named using the ExtendScript convention of `jsxinc` for
scripts that are included into the main file. The script has a number of defaults but exposes them via optional parameters.

The assumption right now is that
* paragraph styles should have their `appliedFont` set to `null` and their
   `basedOn` value set to the base (`DefaultParagraph`) style
* character styles should have their `appliedFont` set to `null` and their
   `basedOn` value set to the first character style, which should be `[None]`
* paragraph styles *may* be grouped, character styles are not

This script can easily be tested by invoking it with `app.activeDocument`
(while a doc is open in InDesign) and "Adobe Arabic" (or another font).