DT470: Switch to `localdb` for PreviewJobs
The only major callout on this functionality is that as
we're using EF Core, we must ensure that the databse
is created rather than using a Code First approach to
handle that for us.

The caveat is that if the model changes, we will need
to delete the previous database manually.
This can be done by deleting the DB files
from the users' home directory, or by using
`SqlLocalDb.exe delete`.

Side Effects:
* Requires installation of SQLExpress
* Change the DB name in `serviceSettings.json`