DT-27: Move document control flow logic into ScriptRunner.cs
We had all of the control flow logic in the InDesign Server scripts and
this meant that we could not cancel or affect the flow once the job was
kicked off.

The flow has been moved into the server, with the ScriptRunner
determining which files should become documents and which documents
should become books.

This ultimately gives us more control over how jobs are executed and
cancelled.